pyproject.toml
[tool.poetry]
name = "shardctl"
version = "0.1.0"
description = "A CLI tool for managing microservices with docker-compose"
authors = ["Pyrofex <dev@pyrofex.net>"]
readme = "README.md"
packages = [{include = "shardctl"}]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.9.0"
rich = "^13.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
shardctl = "shardctl.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
